change ClientService and ExpertService

+++ src/main/java/sa/aref/service/accounts/ClientService.java
@@ -16,7 +16,7 @@ public class ClientService {
         return clientRepository.save(client);
     }
 
-    public void changePassword(Long id, String password) {
+    public void changePassword(Integer id, String password) {
         clientRepository.changePassword(id, password);
     }
